fix : 가게 CRUD 기능 수정

+++ src/main/java/com/team2/grabtable/domain/store/repository/StoreRepository.java
@@ -7,12 +7,12 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface StoreRepository extends JpaRepository<Store, Integer> {
+public interface StoreRepository extends JpaRepository<Store, Long> {
 
     @Query("select s from Store s where s.owner.ownerId = :ownerId")
-    List<Store> findByOwnerId(@Param("ownerId") int ownerId);
+    List<Store> findByOwnerId(@Param("ownerId") Long ownerId);
 
     @Query("select count(s) from Store s where s.owner.ownerId = :ownerId")
-    Long countByOwnerId(@Param("ownerId") int ownerId);
+    Long countByOwnerId(@Param("ownerId") Long ownerId);
 
 }
